Fix editor
