+++ package.json
@@ -12,23 +12,6 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.7.15",
-    "@milkdown/core": "^7.2.3",
-    "@milkdown/ctx": "^7.2.3",
-    "@milkdown/plugin-clipboard": "^7.2.3",
-    "@milkdown/plugin-cursor": "^7.2.3",
-    "@milkdown/plugin-history": "^7.2.3",
-    "@milkdown/plugin-listener": "^7.2.3",
-    "@milkdown/plugin-math": "^7.2.3",
-    "@milkdown/plugin-prism": "^7.2.3",
-    "@milkdown/plugin-tooltip": "^7.2.3",
-    "@milkdown/plugin-trailing": "^7.2.3",
-    "@milkdown/preset-commonmark": "^7.2.3",
-    "@milkdown/preset-gfm": "^7.2.3",
-    "@milkdown/prose": "^7.2.3",
-    "@milkdown/react": "^7.2.3",
-    "@milkdown/theme-nord": "^7.2.3",
-    "@milkdown/transformer": "^7.2.3",
-    "@milkdown/utils": "^7.2.3",
     "@prosemirror-adapter/react": "^0.2.5",
     "@react-pdf/renderer": "^3.1.12",
     "@testing-library/cypress": "^9.0.0",
